move mysql into subfolder
