+++ drivers/mysql/helpers.go
@@ -1,4 +1,4 @@
-package drivers
+package mysql
 
 import (
 	"strconv"
