+++ drivers/mysql/helpers_test.go
@@ -1,4 +1,4 @@
-package drivers
+package mysql
 
 import (
 	"reflect"
